Update meta data to use more correct phrasing
Unfortunately GitHub pages do not keep original update/edit data for files and overwrites them to use the date of the last website deployment. Originally I was using this `meta` line to show the last edit date for the current article specifically, so I've updated the phrasing to be more explicit.